categories model routes controller added

+++ backend/src/controllers/blog.ts
@@ -1,29 +1,29 @@
 import express from "express";
 import { getBlogById, getBlogs, deleteBlogById, createBlog } from "../models/schemas/blog"
-// import { getBlogDetailByBlogId, deleteBlogDetailById, updateBlogDetailById } from "../models/schemas/blogDetail"
-import NotFoundError from "../errors/NotFoundError"
 import { redisClient } from "../config/redis";
+import Response from "../utils/classes/Response";
+import CustomError from "../utils/classes/CustomError";
+import { CACHE_CONFIG } from "../utils/enum";
+import isValidMongoId from "../utils/functions/isMongoId"
+import { getCategoryByName, getCategoryById } from "../models/schemas/category";
+import checkUserPermission from "../utils/functions/checkUserPermission";
 
-import Enum from "../utils/enum";
+const cacheKey = CACHE_CONFIG.BLOGS.KEY
+const cacheExpireTime = CACHE_CONFIG.BLOGS.EXPIRE_TIME
 
 export const getAllBlogs = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
-    const cacheKey = Enum.CACHE_KEYS.BLOGS;
     try {
-        let blogs: any = await redisClient.get(cacheKey);
-        if (blogs) {
+        let cachedBlogs = await redisClient.get(cacheKey);
+        if (cachedBlogs) {
             console.log("cacheden cekti")
-            return res.status(200).json(JSON.parse(blogs));
+            return res.status(200).json(Response.successResponse(JSON.parse(cachedBlogs)));
         }
-        else if (!blogs) {
-            blogs = await getBlogs();
-            if (!blogs) {
-                throw new NotFoundError("Blog bulunamadi");
-            }
-            console.log("veritabanindan cekti")
-            await redisClient.setEx(cacheKey, 600, JSON.stringify(blogs)); // 600 second
+        const blogs = await getBlogs();
+        if (!blogs || blogs.length === 0) {
+            throw new CustomError(404, "Not found", "Bloglar bulunamadı.")
         }
-        // const blogs = await getBlogs();
-        return res.status(200).json(blogs);
+        await redisClient.setEx(cacheKey, cacheExpireTime, JSON.stringify(blogs));
+        return res.status(200).json(Response.successResponse(blogs));
     } catch (error) {
         next(error);
     }
@@ -32,35 +32,42 @@ export const getAllBlogs = async (req: express.Request, res: express.Response, n
 export const getBlogDetail = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
     const blogId = req.params.blogId;
     try {
+        if (!isValidMongoId(blogId)) {
+            throw new CustomError(400, "Validation error", "Invalid blog id");
+        }
         //validation
         const blog = await getBlogById(blogId);
         if (!blog) {
-            throw new NotFoundError("Blog detayi bulunamadi.");
+            throw new CustomError(404, "Not found", "Blog detay bulunamadı.");
         }
-        return res.status(200).json(blog);
+        return res.status(200).json(Response.successResponse(blog));
     } catch (error) {
         next(error);
     }
 }
 
 export const postCreateBlog = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
-    const { title, description, categoryId } = req.body;
+    const { title, description, categoryName } = req.body;
     const userId = res.locals.userId;
-    const cacheKey = Enum.CACHE_KEYS.BLOGS;
     const imageName = req.file ? req.file.filename : null; // burada yalnızca dosya ismini alıyor ve kaydediyor veri tabanına
     try {
-        //validation
+        //validation imageName'i validate et 
+        const category = await getCategoryByName(categoryName);
+        if (!category) {
+            throw new CustomError(400, "Validation error", "Bu isimde bir kategori bulunmuyor.")
+        }
 
         const imageUrl = `${req.protocol}://${req.get('host')}/uploads/${imageName}`;
+
         await createBlog({
             title,
-            imageUrl:imageUrl,
+            imageUrl,
             description,
-            categoryId,
+            categoryId: category._id,
             userId
         })
         await redisClient.del(cacheKey)
-        return res.status(201).json("Blog olusturuldu");
+        return res.status(201).json(Response.successResponse(null, "Blog oluşturuldu."));
     } catch (error) {
         next(error);
     }
@@ -70,15 +77,21 @@ export const deleteBlog = async (req: express.Request, res: express.Response, ne
     const blogId = req.params.blogId;
     const userId = res.locals.userId;
     try {
+        if (!isValidMongoId(blogId)) {
+            throw new CustomError(400, "Validation error", "Invalid blog id");
+        }
         //validation
         const blog = await getBlogById(blogId);
         if (!blog) {
-            throw new NotFoundError("Blog bulunamadi");
+            throw new CustomError(404, "Not found", "Blog bulunamadı.")
+        }
+        console.log(blog.userId?.toString());
+        if (!checkUserPermission(userId, blog.userId?.toString())) {
+            throw new CustomError(401,"Unauthorized","Bu blogu silmeye yetkin yok.")
         }
         await deleteBlogById(blogId);
-        // await deleteBlogDetailById();
-
-        return res.status(201).json("Blog olusturuldu");
+        await redisClient.del(cacheKey)
+        return res.status(201).json(Response.successResponse(null, "Blog başarıyla silindi"));
     } catch (error) {
         next(error);
     }
